requirement.txt
fastapi 
uvicorn 
pillow 
python-multipart 
ultralytics
